Add files via upload
Separated scheduler from dispatcher. The new process will always be added to the ready queue at the appropriate position. If all the tasks are created os_start is set 1. If a task is added to the head then d_flag is set 1. Only when both are set the dispatcher(context switch ) is called